flask_app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.1
